+++ .gitignore
@@ -1,3 +1,5 @@
 .idea
 __pycache__
 .ipynb_checkpoints
+.bash_history
+.sudo_as_admin_successful
