+++ Dockerfile
@@ -23,19 +23,6 @@ RUN apt-get install -y \
         curl \
         graphviz
 
-# not needed since we do not compile from source
-#clang \
-#gcc \
-#g++ \
-#g++-multilib \
-#gdb \
-#clang-tidy \
-#clang-format \
-#gcovr \
-#llvm \
-#libgraphviz-dev \
-#libboost-all-dev \
-
 # This adds the 'default' user to sudoers with full privileges:
 RUN HOME=/home/default && \
     mkdir -p ${HOME} && \
@@ -61,22 +48,9 @@ WORKDIR /build
 
 ARG GIT_REF=main
 
-# Clone and build Lydia
-#RUN git clone --recursive https://github.com/whitemech/lydia.git /build/lydia
-#RUN git checkout ${GIT_REF} &&\
-#    rm -rf build &&\
-#    mkdir build &&\
-#    cd build &&\
-#    cmake -DCMAKE_BUILD_TYPE=Release .. &&\
-#    cmake --build . --target lydia-bin -j4 &&\
-#    make install &&\
-#    cd .. &&\
-#    rm -rf /build/lydia
-
 WORKDIR /home/default
 
 USER default
-
 
 # Install CUDD
 RUN wget https://github.com/whitemech/cudd/releases/download/v${CUDD_VERSION}/cudd_${CUDD_VERSION}_linux-amd64.tar.gz &&\
@@ -94,31 +68,21 @@ RUN wget https://github.com/whitemech/MONA/releases/download/v${MONA_VERSION}/mo
     sudo cp -Pr include/* /usr/local/include &&\
     rm -rf mona_${MONA_VERSION}_linux-amd64*
 
-# Build and install Syft
-#RUN git clone https://github.com/whitemech/Syft.git &&\
-#    cd Syft &&\
-#    git checkout ${SYFT_TAG} &&\
-#    mkdir build && cd build &&\
-#    cmake -DCMAKE_BUILD_TYPE=Release .. &&\
-#    make -j &&\
-#    make install &&\
-#    cd .. &&\
-#    rm -rf Syft
-
 # install Syft
 RUN wget https://github.com/whitemech/Syft/releases/download/v0.1.0/syft-3.0.0_ubuntu-20.04.tar.gz && \
-  tar -xf syft-3.0.0_ubuntu-20.04.tar.gz &&\
-  cd syft-3.0.0_ubuntu-20.04 && \
-  sudo cp -P lib/* /usr/local/lib/
+    tar -xf syft-3.0.0_ubuntu-20.04.tar.gz &&\
+    cd syft-3.0.0_ubuntu-20.04 && \
+    sudo cp -P lib/* /usr/local/lib/
 
 # install Lydia
 RUN wget https://github.com/whitemech/lydia/releases/download/v0.1.3/lydia &&\
-  chmod u+x lydia &&\
-  sudo cp lydia /usr/local/bin/lydia &&\
-  sudo chown default:default /usr/local/bin/lydia
+    chmod u+x lydia &&\
+    sudo cp lydia /usr/local/bin/lydia &&\
+    sudo chown default:default /usr/local/bin/lydia
 
+# install MDP-Planner-Chip-Production
+RUN git clone https://github.com/iaiamomo/MDP-Planner-Chip-Production.git &&\
+    cd MDP-Planner-Chip-Production &&\
+    pip install -e .
 
-
-
-# RUN pip install --no-cache-dir -r /home/default/MDP-Planner-Chip-Production/requirements.txt
-RUN cd MDP-Planner-Chip-Production && pip install -e .
+WORKDIR /home/default/MDP-Planner-Chip-Production
