+++ docs/notebooks/config.json
@@ -1,0 +1,5 @@
+{
+    "mode": "automata",
+    "phase": 2,
+    "size": "small"
+}
