+++ docs/notebooks/experimental_results/memory_profiler_composition_automata_phase2_small.log
@@ -1,0 +1,10 @@
+Filename: /home/default/MDP-Planner-Chip-Production/docs/notebooks/chip_production_supply_chain.py
+
+Line #    Mem usage    Increment  Occurrences   Line Contents
+=============================================================
+    20    131.0 MiB    131.0 MiB           1   @profile(stream=fp_compMDP)
+    21                                         def execute_composition_automata(target, services):
+    22    264.6 MiB    133.6 MiB           1       mdp = composition_mdp(target, *services, gamma=0.9)
+    23    264.6 MiB      0.0 MiB           1       return mdp
+
+
