+++ docs/notebooks/experimental_results/memory_profiler_policy_automata_phase2_small.log
@@ -1,0 +1,11 @@
+Filename: /home/default/MDP-Planner-Chip-Production/docs/notebooks/chip_production_supply_chain.py
+
+Line #    Mem usage    Increment  Occurrences   Line Contents
+=============================================================
+    30    264.6 MiB    264.6 MiB           1   @profile(stream=fp_DPAnalytic)
+    31                                         def execute_policy(mdp):
+    32    264.6 MiB      0.0 MiB           1       opn = DPAnalytic(mdp, 1e-4)
+    33    264.6 MiB      0.0 MiB           1       opt_policy = opn.get_optimal_policy_vi()
+    34    264.6 MiB      0.0 MiB           1       return opt_policy
+
+
