+++ docs/notebooks/experimental_results/time_profiler_automata_phase2_small.txt
@@ -1,0 +1,3 @@
+tot_services: 14
+composition elapsed time: 104.805862022 s
+policy elapsed time: 0.886436142 s
