+++ setup.py
@@ -17,6 +17,8 @@ setup(name='stochastic_service_composition',
             "pythomata",
             "flask",
             "aiohttp",
-            "aiohttp_jinja2"
+            "aiohttp_jinja2",
+            "mdp_dp_rl @ git+https://github.com/luusi/mdp-dp-rl.git#egg=mdp_dp_rl",
+            "memory-profiler"
       ]
       )
