+++ stochastic_service_composition.egg-info/PKG-INFO
@@ -2,8 +2,10 @@ Metadata-Version: 2.1
 Name: stochastic-service-composition
 Version: 0.1.0
 Summary: Implementation of stochastic service composition.
-Home-page: http://github.com/luusi/stochastic-service-composition
-Author: Luciana Silo
-Author-email: silo.1586010@studenti.uniroma1.it
+Home-page: UNKNOWN
 License: MIT
+Platform: UNKNOWN
 License-File: LICENSE
+
+UNKNOWN
+
