+++ stochastic_service_composition.egg-info/requires.txt
@@ -1,12 +1,13 @@
-numpy
-graphviz
-websockets
-paho-mqtt
-requests
-connexion[swagger-ui]
-logaut
-pythomata
-flask
 aiohttp
 aiohttp_jinja2
+connexion[swagger-ui]
+flask
+graphviz
+logaut
 mdp_dp_rl@ git+https://github.com/luusi/mdp-dp-rl.git#egg=mdp_dp_rl
+memory-profiler
+numpy
+paho-mqtt
+pythomata
+requests
+websockets
